+++ examples/example_evaluate_without_formula2.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from brdr.aligner import Aligner
 from brdr.constants import EVALUATION_FIELD_NAME
-from brdr.enums import GRBType, AlignerResultType, Full
+from brdr.enums import GRBType, AlignerResultType, FullStrategy
 from brdr.geometry_utils import geom_from_wkt
 from brdr.grb import GRBActualLoader
 from brdr.loader import DictLoader
@@ -28,7 +28,7 @@ if __name__ == "__main__":
     dict_evaluated, prop_dictionary = actual_aligner.evaluate(
         ids_to_evaluate=None,
         base_formula_field=None,
-        full_strategy=Full.PREFER_FULL,
+        full_strategy=FullStrategy.PREFER_FULL,
         relevant_distances=relevant_distances,
     )