+++ examples/example_multi_to_single_modus.py
@@ -1,5 +1,5 @@
 from brdr.aligner import Aligner
-from brdr.enums import GRBType, OpenbaarDomeinStrategy
+from brdr.enums import GRBType, OpenDomainStrategy
 from brdr.grb import GRBActualLoader
 from brdr.oe import OnroerendErfgoedLoader, OEType
 from examples import print_brdr_formula
@@ -15,7 +15,7 @@ if __name__ == "__main__":
     # Example (ErfgoedObject): https://inventaris.onroerenderfgoed.be/erfgoedobjecten/305858
     loader = OnroerendErfgoedLoader([305858], oetype=OEType.EO)
     relevant_distance = 5  # rd is taken very high to show the difference
-    od_strategy = OpenbaarDomeinStrategy.SNAP_ALL_SIDE
+    od_strategy = OpenDomainStrategy.SNAP_ALL_SIDE
     threshold_circle_ratio = 0.75  # default it is 0.98, but because it are not fully circles in this example we put this on 0.75
 
     # EXAMPLE of "multi_as_single_modus"=FALSE