+++ examples/example_readme.py
@@ -1,5 +1,5 @@
 from brdr.aligner import Aligner
-from brdr.enums import OpenbaarDomeinStrategy
+from brdr.enums import OpenDomainStrategy
 from brdr.geometry_utils import geom_from_wkt
 from brdr.loader import DictLoader
 
@@ -19,7 +19,7 @@ aligner.load_reference_data(loader)
 relevant_distance = 1
 process_result = aligner.process(
     relevant_distance=relevant_distance,
-    od_strategy=OpenbaarDomeinStrategy.SNAP_INNER_SIDE,
+    od_strategy=OpenDomainStrategy.SNAP_INNER_SIDE,
     threshold_overlap_percentage=50,
 )
 # PRINT RESULTS IN WKT